[feature|evaluator] #33 0.3.2.13.1 문자열 평가 기능 추가 및 (미완성) 함수 호출 평가 기능 구현

+++ test/test.cpp
@@ -21,7 +21,7 @@ const int main(const size_t argc, const char* const argv[])
 #endif
 
         std::cout << "options: ";
-		for (int i = 1; i < argc; ++i)
+		for (size_t i = 1; i < argc; ++i)
 		{
 			std::cout << argv[i] << " ";
 #ifdef _WIN32
